start detec script
