+++ simulation_settings.json
@@ -1,28 +1,18 @@
 {
-  "sn_settings": {
-      "tnsname": "2023vbg",
-      "filt": "o", 
-      "num_controls": 8,
-      "mjd0": 60212.497002,
-      "mjd_bin_size": 1.0,
-      "badday_flag": 8388608
+  "sn_info": {
+    "tnsname": "2023vbg",
+    "filt": "o", 
+    "num_controls": 8,
+    "mjd0": 60212.497002,
+    "mjd_bin_size": 1.0,
+    "badday_flag": 8388608
   },
 
   "data_dir": "/Users/sofiarest/Desktop/Supernovae/data_refactor/output/misc/2023vbg",
   "sim_tables_dir": "/Users/sofiarest/Desktop/Supernovae/data_refactor/output/misc/2023vbg/bump_analysis/sim_tables",
   "detec_tables_dir": "/Users/sofiarest/Desktop/Supernovae/data_refactor/output/misc/2023vbg/bump_analysis/detec_tables",
-  
-  "peak_mag_settings": {
-      "peak_mag_min": 23,
-      "peak_mag_max": 16,
-      "n_peaks": 20
-  },
-
-  "num_iterations": 50000,
 
   "skip_control_ix": [],
-
-  "observation_seasons": [],
 
   "sigma_kerns": [5, 20, 40, 80, 100, 150, 200]
 }
